refactor(settings-dialog): extract dialog chrome into a part file
Move the frame builders — loading/error states, header, tab strip, and
footer — out of system_emulator_settings_dialog.dart into a new part file
`system_emulator_settings_dialog/chrome.dart` as `extension _Chrome`.

Pure move, zero behaviour change. State stays on the host State (extensions
can't declare fields); the one `setState` call routes through the existing
host `rebuild` bridge. Host 1802 → 1534.
